It exports, but the export is not yet quite useable

+++ db/migrate/001_patch_projects.rb
@@ -2,7 +2,7 @@ class PatchProjects < ActiveRecord::Migration
   def change
     add_column :projects, :tj_activated, :boolean
     add_column :projects, :tj_version, :string
-    add_column :projects, :tj_dailyworkinghours, :integer
+    add_column :projects, :tj_dailyworkinghours, :float
     add_column :projects, :tj_period,  :string
     add_column :projects, :tj_currency,  :string
     add_column :projects, :tj_now,  :string
